update & fix mina contract.
In the local tests I did before, I realized that I wrote the code requesting the Oracle Node into the mina contract. Due to this error, I moved my Oracle code here to mina contract test code for now. In the current situation, after the request sent to the contract is approved, it pulls the data and sends it to my simple code on my localhost server.