update Oracle
Simply, when the state value in the mina contract is updated, the frontend transmits the data to the oracle. Then, it transmits this value to the module in aptos via the oracle admin wallet.

My current plan is to develop this oracle and have it receive the updates by just listening to the mina contract without receiving data from the frontend.